requirements.txt
# ══════════════════════════════════════════════════════════════
#  Omnium — Python Dependencies
#  Install: pip install -r requirements.txt
# ══════════════════════════════════════════════════════════════

# ── Core ──
flask>=3.0                  # REST API framework
mariadb>=1.1                # MariaDB connector (requires MariaDB C connector)
python-dotenv>=1.0          # Load .env files for local config

# ── Data & ML ──
yfinance>=0.2               # Yahoo Finance — used by seed.py
scikit-learn>=1.4            # ML trading algorithm (LinearRegression)
numpy>=1.26                  # Numerical operations for ML features

# ── Dev & Testing ──
pytest>=7.4
pytest-cov>=4.1
ruff>=0.4                   # Linter

# ── DevTools ──
watchdog>=3.0               # PlantUML file watcher
